Paula

+++ backend/src/main/java/c23_99_m_webapp/backend/services/ResourceService.java
@@ -1,7 +1,8 @@
 package c23_99_m_webapp.backend.services;
 
+import c23_99_m_webapp.backend.exceptions.BadCustomerRequestException;
 import c23_99_m_webapp.backend.exceptions.ResourceNotFoundException;
-import c23_99_m_webapp.backend.mappers.ResourceCreateMapper;
+import c23_99_m_webapp.backend.exceptions.ResourceUnavailableException;
 import c23_99_m_webapp.backend.mappers.ResourceViewMapper;
 import c23_99_m_webapp.backend.models.dtos.ResourceCreateDTO;
 import c23_99_m_webapp.backend.models.dtos.ResourceViewDTO;
@@ -9,21 +10,26 @@ import c23_99_m_webapp.backend.models.Resource;
 import c23_99_m_webapp.backend.models.enums.ResourceStatus;
 import c23_99_m_webapp.backend.repositories.ReservationRepository;
 import c23_99_m_webapp.backend.repositories.ResourceRepository;
+import c23_99_m_webapp.backend.validations.ValidationResource;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static c23_99_m_webapp.backend.mappers.ResourceCreateMapper.toEntity;
 
 @Service
 public class ResourceService {
 
     private final ResourceRepository resourceRepository;
     private final ReservationRepository reservationRepository;
+    private final List<ValidationResource> validations;
 
-    public ResourceService(ResourceRepository resourceRepository, ReservationRepository reservationRepository){
+    public ResourceService(ResourceRepository resourceRepository, ReservationRepository reservationRepository, List<ValidationResource> validations){
         this.resourceRepository = resourceRepository;
         this.reservationRepository = reservationRepository;
-
+        this.validations = validations;
     }
 
     public List<ResourceViewDTO> getResources() {
@@ -32,12 +38,12 @@ public class ResourceService {
 
     }
 
-//    public ResourceViewDTO getResourceById(Long id) {
-//        Resource resource = resourceRepository.findById(id)
-//                .orElseThrow(() -> new ResourceNotFoundException("No se encontró el recurso con Id: " + id));
-//
-//        return ResourceViewMapper.toDTO(resource);
-//    }
+    public ResourceViewDTO getResourceById(Long id) {
+        Resource resource = resourceRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("No se encontró el recurso con Id: " + id));
+
+        return ResourceViewMapper.toDTO(resource);
+    }
 //
 //    public ResourceViewDTO saveResource(ResourceCreateDTO resourceDTO){
 //        Resource resource = resourceRepository.save(toEntity(resourceDTO));
@@ -46,7 +52,10 @@ public class ResourceService {
 //    }
 
     public Resource saveResourceEntity(ResourceCreateDTO resourceDTO) {
-        Resource resource = ResourceCreateMapper.toEntity(resourceDTO);
+        for (ValidationResource v : validations) {
+            v.validate(resourceDTO);
+        }
+        Resource resource = toEntity(resourceDTO);
         return resourceRepository.save(resource);
     }
 
@@ -60,6 +69,9 @@ public class ResourceService {
     public ResourceViewDTO updateResource(long id, ResourceCreateDTO dto) {
         Resource existingResource = resourceRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("No se encontró el recurso con Id: " + id));
+        for (ValidationResource v : validations) {
+            v.validate(dto);
+        }
 
         existingResource.setName(dto.name());
         existingResource.setDescription(dto.description());
@@ -71,20 +83,21 @@ public class ResourceService {
         return ResourceViewMapper.toDTO(modifiedResource);
     }
 
-    public void deleteResource(long id) {
+    @Transactional
+    public void deleteResource(long id){
         Resource resource = resourceRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("No se encontro al recurso con id: " + id));
 
         boolean existingResourceReserve = this.reservationRepository.existsByResource(resource);
 
-//        if(existingResourceReserve){
-//            throw new BadCustomerRequestException("No es posible eliminar un recurso con reservas asociadas");
-//        }
+        if(existingResourceReserve){
+            throw new BadCustomerRequestException("No es posible eliminar un recurso con reservas asociadas");
+        }
         this.resourceRepository.deleteById(id);
     }
 
 
-    public ResourceViewDTO updateResourceStatus(Long id, ResourceStatus status){
+    public ResourceViewDTO updateAndReturnResourceStatus(Long id, ResourceStatus status){
         Resource resource = resourceRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("No se encontró el recurso con Id: " + id));
 
@@ -98,6 +111,22 @@ public class ResourceService {
     public List<ResourceViewDTO> getResourcesByStatus(ResourceStatus status){
         List<Resource> resourcesByStatus = resourceRepository.findAllByStatus(status);
         return resourcesByStatus.stream().map(ResourceViewMapper::toDTO).collect(Collectors.toList());
+    }
+
+    public void validateResourceAvailability(Long resourceId) {
+        Resource resource = resourceRepository.findById(resourceId)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso no encontrado con ID: " + resourceId));
+
+        if (!resource.getStatus().equals(ResourceStatus.AVAILABLE)) {
+            throw new ResourceUnavailableException("El recurso no está disponible para reservar.");
+        }
+    }
+
+    public void updateResourceStatus(Long resourceId, ResourceStatus status) {
+        Resource resource = resourceRepository.findById(resourceId)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso no encontrado con ID: " + resourceId));
+        resource.setStatus(status);
+        resourceRepository.save(resource);
     }
